Refactor for tutorial
