+++ README.md
@@ -2,7 +2,7 @@
 
 Is a learning project for a blog post.
 
-Technlogies:
+Technologies:
 
 * Azure App Service
 * Azure CosmosDB
@@ -11,5 +11,5 @@ Technlogies:
 
 Purpose:
 
-It is a small application that provides some basic functionalitty to check the site health every N minutes.
+It is a small application that provides some basic functionality to check the site health every N minutes.
 It would then send an SMS alert to a configured cell phone number in case the site goes offline.
