--- SiteCheckRobot.WebApi/Program.cs
@@ -1,53 +1,0 @@
-using Azure.Extensions.AspNetCore.Configuration.Secrets;
-using Azure.Identity;
-using Azure.Security.KeyVault.Secrets;
-using Microsoft.Extensions.Configuration;
-using SiteCheckRobot.Core;
-
-namespace SiteCheckRobot.WebApi
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var builder = WebApplication.CreateBuilder(args);
-
-            // Add services to the container.
-
-            builder.Services.AddControllers();
-            // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-            builder.Services.AddEndpointsApiExplorer();
-            builder.Services.AddSwaggerGen();
-
-            // add background web worker
-            builder.Services.AddHostedService<SiteCheckBackgroundWorker>();
-
-            builder.Services.AddTransient<IPeriodicWork, PeriodicWork>();
-
-            var keyVaultUrl = builder.Configuration["keyVaultUrl"];
-            var secretClient = new SecretClient(new Uri(keyVaultUrl), new DefaultAzureCredential());
-            builder.Configuration.AddAzureKeyVault(secretClient, new KeyVaultSecretManager());
-            
-            var app = builder.Build();
-
-            // Configure the HTTP request pipeline.
-            if (app.Environment.IsDevelopment())
-            {
-                app.UseSwagger();
-                app.UseSwaggerUI();
-            }
-
-            app.UseHttpsRedirection();
-
-            app.UseAuthorization();
-
-
-            app.MapControllers();
-            
-            
-            
-            app.Run();
-        }
-    }
-
-}
