+++ SiteCheckRobot/Properties/launchSettings.json
@@ -16,9 +16,7 @@
       "launchUrl": "swagger",
       "applicationUrl": "https://localhost:7231;http://localhost:5075",
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "VaultUri": "https://sitecheckrobotvault.vault.azure.net/",
-        "AZURE_USERNAME": "matthias.jost@gmail.com"
+        "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
     "IIS Express": {
@@ -26,9 +24,7 @@
       "launchBrowser": true,
       "launchUrl": "swagger",
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development",
-        "VaultUri": "https://sitecheckrobotvault.vault.azure.net/",
-        "AZURE_USERNAME": "matthias.jost@gmail.com"
+        "ASPNETCORE_ENVIRONMENT": "Development"
       }
     }
   }
