+++ SiteCheckRobot/SiteCheckRobot.csproj
@@ -13,10 +13,7 @@
     <PackageReference Include="Microsoft.Azure.Cosmos" Version="3.21.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Cosmos" Version="6.0.1" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SiteCheckRobot.Core\SiteCheckRobot.Core.csproj" />
+    <PackageReference Include="Twilio" Version="6.4.0" />
   </ItemGroup>
 
 </Project>
